+++ docs/.vitepress/components/FramePostLayout.vue
@@ -27,7 +27,7 @@ onMounted(async () => {
 <!-- 导入主题样式 -->
 <style src="../../../src/templates/components/theme-dark/styles.less" scoped></style>
 <style src="../../../src/templates/components/theme-light/styles.less" scoped></style>
-<style src="../../../src/templates/page-components/utils/main.css" scoped></style>
+<style src="../../../src/templates/page-components/shared/main.css" scoped></style>
 <style src="../../../src/templates/components/text-size-normal/styles.css" scoped></style>
 <!-- 导入 post 页样式 -->
 <style src="../../../src/templates/page-components/post/styles.css" scoped></style>