+++ src/templates/page-components/links/index.ts
@@ -1,3 +1,3 @@
-import "../utils/main";
+import "../shared/main";
 import "./styles.css";
 import "../utils/article.css";