+++ src/templates/page-components/post/index.ts
@@ -1,4 +1,4 @@
-import "../utils/main";
+import "../shared/main";
 import "./styles.css";
 import "../utils/article.css";
 import "../utils/article-metadata.css";